+++ src/commands/delegates.ts
@@ -6,6 +6,8 @@ export type ServiceMapCondition<T, V> = (item: T, index: number) => V;
 
 export type CompareCondition<T> = (first: T, second: T) => number;
 
+export type EqualityCondition<T> = (first: T, second: T) => boolean;
+
 export type ReduceCondition<T> = (first: T, second: T) => T;
 export type ReduceWithAccumulatorCondition<T, V> = (acc: V, second: T) => V;